Add files via upload
This project is a Tic-Tac-Toe game of 3 by 3 board for 2 players where each player has 3 identical signs different from that of the other player (i.e. 3 X's and 3 O's) to be placed on the game board. For a winner to emerge, he has to place his signs in turn on the same column, row or diagonal of the game board. Note that 3 empty spaces will be left out after the players must have placed their individual signs on the game board in turn.
After the placement of the signs and no winner emerged, the players have to be moving their individual signs in turn  on the game board until either player's signs fall on the same column, row or diagonal thus, a winner emerges.